refactor: move `Timing` to shared
