+++ .vscode/settings.json
@@ -42,6 +42,7 @@
     "Minifier",
     "outro",
     "Paneview",
+    "pathe",
     "picocolors",
     "poolifier",
     "preact",
