+++ packages/ast-utils/package.json
@@ -22,7 +22,6 @@
     "./program": "./src/program.ts",
     "./reference": "./src/reference.ts",
     "./scope": "./src/scope.ts",
-    "./timing": "./src/timing.ts",
     "./types": "./src/types.ts"
   },
   "files": [
