+++ packages/shared/src/timing.ts
