+++ packages/unminify/package.json
@@ -30,6 +30,7 @@
     "@babel/preset-env": "^7.23.3",
     "@babel/types": "^7.23.4",
     "lebab": "^3.2.3",
+    "pathe": "^1.1.1",
     "picocolors": "^1.0.0",
     "prettier": "^2.8.8",
     "zod": "^3.22.4"
